movie_finder_app/requirements.txt
python-dotenv
requests
flask
gunicorn
pre-commit
